move build_qlateral_array function from AbstractNetwork to individual networks
