+++ src/troute-network/troute/AbstractNetwork.py
@@ -1,14 +1,10 @@
 from abc import ABC, abstractmethod
 from functools import partial
 import pandas as pd
-import numpy as np
-import pyarrow.parquet as pq
 from datetime import datetime, timedelta
-from joblib import delayed, Parallel
-import netCDF4
+
 import time
 import logging
-import pathlib
 
 from troute.nhd_network import extract_connections, replace_waterbodies_connections, reverse_network, reachable_network, split_at_waterbodies_and_junctions, split_at_junction, dfs_decomposition
 from troute.nhd_network_utilities_v02 import organize_independent_networks, build_channel_initial_state, build_refac_connections
@@ -27,7 +23,8 @@ class AbstractNetwork(ABC):
                 "_reverse_network", "_q0", "_t0", "_link_lake_crosswalk",
                 "_qlateral", "_break_segments", "_segment_index",
                 "supernetwork_parameters", "waterbody_parameters","data_assimilation_parameters",
-                "restart_parameters", "compute_parameters", "verbose", "showtiming", "break_points"]
+                "restart_parameters", "compute_parameters", "forcing_parameters",
+                "hybrid_parameters", "verbose", "showtiming", "break_points"]
     
     def __init__(self,):
 
@@ -55,15 +52,13 @@ class AbstractNetwork(ABC):
                 self._break_segments = self._break_segments | set(self.waterbody_connections.values())
             if self.break_points["break_network_at_gages"]:
                 self._break_segments = self._break_segments | set(self.gages.get('gages').keys())
-        
-        #self.build_diffusive_domain(compute_parameters)
 
         self.create_independent_networks()
 
         self.initial_warmstate_preprocess()
 
 
-    def assemble_forcings(self, run, forcing_parameters, hybrid_parameters, supernetwork_parameters, cpu_pool):
+    def assemble_forcings(self, run,):
         """
         Assemble model forcings. Forcings include hydrological lateral inflows (qlats)
         and coastal boundary depths for hybrid runs
@@ -91,13 +86,13 @@ class AbstractNetwork(ABC):
         """
     
         # Unpack user-specified forcing parameters
-        dt                           = forcing_parameters.get("dt", None)
-        qts_subdivisions             = forcing_parameters.get("qts_subdivisions", None)
-        qlat_input_folder            = forcing_parameters.get("qlat_input_folder", None)
-        qlat_file_index_col          = forcing_parameters.get("qlat_file_index_col", "feature_id")
-        qlat_file_value_col          = forcing_parameters.get("qlat_file_value_col", "q_lateral")
-        qlat_file_gw_bucket_flux_col = forcing_parameters.get("qlat_file_gw_bucket_flux_col", "qBucket")
-        qlat_file_terrain_runoff_col = forcing_parameters.get("qlat_file_terrain_runoff_col", "qSfcLatRunoff")
+        dt                           = self.forcing_parameters.get("dt", None)
+        qts_subdivisions             = self.forcing_parameters.get("qts_subdivisions", None)
+        qlat_input_folder            = self.forcing_parameters.get("qlat_input_folder", None)
+        qlat_file_index_col          = self.forcing_parameters.get("qlat_file_index_col", "feature_id")
+        qlat_file_value_col          = self.forcing_parameters.get("qlat_file_value_col", "q_lateral")
+        qlat_file_gw_bucket_flux_col = self.forcing_parameters.get("qlat_file_gw_bucket_flux_col", "qBucket")
+        qlat_file_terrain_runoff_col = self.forcing_parameters.get("qlat_file_terrain_runoff_col", "qSfcLatRunoff")
 
     
         # TODO: find a better way to deal with these defaults and overrides.
@@ -117,21 +112,27 @@ class AbstractNetwork(ABC):
 
         # Place holder, if reading qlats from a file use this.
         # TODO: add an option for reading qlat data from BMI/model engine
+        start_time = time.time()
+        LOG.info("Creating a DataFrame of lateral inflow forcings ...")
+
         from_file = True
         if from_file:
             self.build_qlateral_array(
                 run,
-                cpu_pool,
-                supernetwork_parameters,
             )
+        
+        LOG.debug(
+            "lateral inflow DataFrame creation complete in %s seconds." \
+                % (time.time() - start_time)
+                )
 
         #---------------------------------------------------------------------
         # Assemble coastal coupling data [WIP]
         #---------------------------------------------------------------------
         # Run if coastal_boundary_depth_df has not already been created:
         if self._coastal_boundary_depth_df.empty:
-            coastal_boundary_elev_files = forcing_parameters.get('coastal_boundary_input_file', None) 
-            coastal_boundary_domain_files = hybrid_parameters.get('coastal_boundary_domain', None)    
+            coastal_boundary_elev_files = self.forcing_parameters.get('coastal_boundary_input_file', None) 
+            coastal_boundary_domain_files = self.hybrid_parameters.get('coastal_boundary_domain', None)    
             
             if coastal_boundary_elev_files:
                 #start_time = time.time()    
@@ -405,213 +406,6 @@ class AbstractNetwork(ABC):
         else:
             self._dataframe = self._dataframe.astype(type)
             
-    
-    def build_diffusive_domain(self, compute_parameters):
-        """
-        
-        """
-        hybrid_params = compute_parameters.get("hybrid_parameters", False)
-        if hybrid_params:
-            # switch parameters
-            # if run_hybrid = False, run MC only
-            # if run_hybrid = True, if use_topobathy = False, run MC+diffusive on RouteLink.nc
-            #    "      "      "  , if use_topobathy = True,  if run_refactored_network = False, run MC+diffusive on original hydrofabric
-            #    "      "      "  , if use_topobathy = True,  if run_refactored_network = True,  run MC+diffusive on refactored hydrofabric
-            run_hybrid             = hybrid_params.get('run_hybrid_routing', False)
-            use_topobathy          = hybrid_params.get('use_natl_xsections', False) 
-            run_refactored         = hybrid_params.get('run_refactored_network', False)
-            
-            # file path parameters of non-refactored hydrofabric defined by RouteLink.nc
-            domain_file            = hybrid_params.get("diffusive_domain",   None)
-            topobathy_file         = hybrid_params.get("topobathy_domain",   None)
-            
-            # file path parameters of refactored hydrofabric for diffusive wave channel routing 
-            refactored_domain_file    = hybrid_params.get("refactored_domain", None)
-            refactored_topobathy_file = hybrid_params.get("refactored_topobathy_domain", None)
-            #-------------------------------------------------------------------------
-            # for non-refactored hydofabric defined by RouteLink.nc
-            # TODO: By default, make diffusive available for both non-refactored and refactored hydrofabric for now. Place a switch in the future. 
-            if run_hybrid and domain_file:
-                
-                LOG.info('reading diffusive domain extent for MC/Diffusive hybrid simulation')
-                
-                # read diffusive domain dictionary from yaml or json
-                diffusive_domain = nhd_io.read_diffusive_domain(domain_file)           
-                
-                if use_topobathy and topobathy_file:
-                    
-                    LOG.debug('Natural cross section data on original hydrofabric are provided.')
-                    
-                    # read topobathy domain netcdf file, set index to 'comid'
-                    # TODO: replace 'link' with a user-specified indexing variable name.
-                    # ... if for whatever reason there is not a `link` variable in the 
-                    # ... dataframe returned from read_netcdf, then the code would break here.
-                    self._topobathy_df = (nhd_io.read_netcdf(topobathy_file).set_index('link'))
-                    
-                    # TODO: Request GID make comID variable an integer in their product, so
-                    # we do not need to change variable types, here.
-                    self._topobathy_df.index = self._topobathy_df.index.astype(int)
-                    
-                else:
-                    self._topobathy_df = pd.DataFrame()
-                    LOG.debug('No natural cross section topobathy data provided. Hybrid simualtion will run on compound trapezoidal geometry.')
-                
-                # initialize a dictionary to hold network data for each of the diffusive domains
-                self._diffusive_network_data = {}
-            
-            else:
-                diffusive_domain       = None
-                self._diffusive_network_data = None
-                self._topobathy_df     = pd.DataFrame()
-                LOG.info('No diffusive domain file specified in configuration file. This is an MC-only simulation')
-            self._unrefactored_topobathy_df  = pd.DataFrame()    
-            #-------------------------------------------------------------------------
-            # for refactored hydofabric 
-            if run_hybrid and run_refactored and refactored_domain_file:
-                
-                LOG.info('reading refactored diffusive domain extent for MC/Diffusive hybrid simulation')
-                
-                # read diffusive domain dictionary from yaml or json
-                self._refactored_diffusive_domain = nhd_io.read_diffusive_domain(refactored_domain_file)           
-                
-                if use_topobathy and refactored_topobathy_file:
-                    
-                    LOG.debug('Natural cross section data of refactored hydrofabric are provided.')
-                    
-                    # read topobathy domain netcdf file, set index to 'comid'
-                    # TODO: replace 'link' with a user-specified indexing variable name.
-                    # ... if for whatever reason there is not a `link` variable in the 
-                    # ... dataframe returned from read_netcdf, then the code would break here.
-                    self._topobathy_df = (nhd_io.read_netcdf(refactored_topobathy_file).set_index('link'))
-
-                    # unrefactored_topobaty_data is passed to diffusive kernel to provide thalweg elevation of unrefactored topobathy 
-                    # for crosswalking water elevations between non-refactored and refactored hydrofabrics. 
-                    self._unrefactored_topobathy_df       = (nhd_io.read_netcdf(topobathy_file).set_index('link'))
-                    self._unrefactored_topobathy_df.index = self._unrefactored_topobathy_df.index.astype(int)
-                    
-                else:
-                    self._topobathy_df = pd.DataFrame()
-                    LOG.debug('No natural cross section topobathy data of refactored hydrofabric provided. Hybrid simualtion will run on compound trapezoidal geometry.')
-                
-                # initialize a dictionary to hold network data for each of the diffusive domains
-                refactored_diffusive_network_data = {}
-            
-            else:
-                self._refactored_diffusive_domain       = None
-                refactored_diffusive_network_data = None
-                self._refactored_reaches                = {}
-                LOG.info('No refactored diffusive domain file specified in configuration file. This is an MC-only simulation')     
-    
-        else:
-            diffusive_domain                  = None
-            self._diffusive_network_data            = None
-            self._topobathy_df                = pd.DataFrame()
-            self._unrefactored_topobathy_df         = pd.DataFrame() 
-            self._refactored_diffusive_domain       = None
-            refactored_diffusive_network_data = None   
-            self._refactored_reaches                = {}
-            LOG.info('No hybrid parameters specified in configuration file. This is an MC-only simulation')
-        
-        #============================================================================
-        # build diffusive domain data and edit MC domain data for hybrid simulation
-        
-        #
-        if diffusive_domain:
-            rconn_diff0 = reverse_network(self._connections)
-            self._refactored_reaches = {}
-            
-            for tw in diffusive_domain:
-                mainstem_segs = diffusive_domain[tw]['links']
-                # we want mainstem_segs start at a mainstem link right after the upstream boundary mainstem link, which is
-                # in turn not under any waterbody. This boundary mainstem link should be turned into a tributary segment.
-                upstream_boundary_mainstem_link = diffusive_domain[tw]['upstream_boundary_link_mainstem']         
-                if upstream_boundary_mainstem_link[0] in mainstem_segs:
-                    mainstem_segs.remove(upstream_boundary_mainstem_link[0])
-                
-                # ===== build diffusive network data objects ==== 
-                self._diffusive_network_data[tw] = {}
-
-                # add diffusive domain segments
-                self._diffusive_network_data[tw]['mainstem_segs'] =  mainstem_segs
-
-                # diffusive domain tributary segments
-                trib_segs = []
-                
-                for seg in mainstem_segs:
-                    us_list = rconn_diff0[seg]
-                    for u in us_list:
-                        if u not in mainstem_segs:
-                            trib_segs.append(u) 
-
-                self._diffusive_network_data[tw]['tributary_segments'] = trib_segs
-                # diffusive domain connections object
-                self._diffusive_network_data[tw]['connections'] = {k: self._connections[k] for k in (mainstem_segs + trib_segs)}       
-
-                # diffusive domain reaches and upstream connections. 
-                # break network at tributary segments
-                _, reaches, rconn_diff = organize_independent_networks(
-                    self._diffusive_network_data[tw]['connections'],
-                    set(trib_segs),
-                    set(),
-                )
-                
-                self._diffusive_network_data[tw]['rconn'] = rconn_diff
-                self._diffusive_network_data[tw]['reaches'] = reaches[tw]
-
-                # RouteLink parameters
-                self._diffusive_network_data[tw]['param_df'] = self._dataframe.filter(
-                    (mainstem_segs + trib_segs),
-                    axis = 0,
-                )
-                self._diffusive_network_data[tw]['upstream_boundary_link'] = upstream_boundary_mainstem_link
-
-                if self._refactored_diffusive_domain: 
-                    diffusive_parameters = {'geo_file_path': refactored_topobathy_file}
-                    refactored_connections = build_refac_connections(diffusive_parameters)
-
-                    # list of stream segments of a single refactored diffusive domain 
-                    refac_tw = self._refactored_diffusive_domain[tw]['refac_tw']
-                    rlinks_tw = self._refactored_diffusive_domain[tw]['rlinks']
-                    refactored_connections_tw = {}   
-
-                    # Subset a connection dictionary (upstream segment as key : downstream segments as values) from refactored_connections
-                    # for a single refactored diffusive domain defined by a current tw. 
-                    for k in rlinks_tw:
-                        if k in refactored_connections.keys() and k != refac_tw:
-                            refactored_connections_tw[k] = refactored_connections[k]
-
-                    refactored_diffusive_network_data[refac_tw] = {}                
-                    refactored_diffusive_network_data[refac_tw]['tributary_segments'] = trib_segs
-                    refactored_diffusive_network_data[refac_tw]['connections'] = refactored_connections_tw                 
-
-                    for k in trib_segs:
-                        refactored_diffusive_network_data[refac_tw]['connections'][k]= [self._refactored_diffusive_domain[tw]['incoming_tribs'][k]]
-
-                    # diffusive domain reaches and upstream connections. 
-                    # break network at tributary segments
-                    _, refactored_reaches_batch, refactored_conn_diff = organize_independent_networks(
-                                                                refactored_diffusive_network_data[refac_tw]['connections'],
-                                                                set(trib_segs),
-                                                                set(),
-                                                                )
-
-                    self._refactored_reaches[refac_tw] = refactored_reaches_batch[refac_tw]
-                    refactored_diffusive_network_data[refac_tw]['mainstem_segs'] = self._refactored_diffusive_domain[tw]['rlinks']
-                    refactored_diffusive_network_data[refac_tw]['upstream_boundary_link'] = self._diffusive_network_data[tw]['upstream_boundary_link'] 
-                else:
-                    self._refactored_reaches={}
-
-                # ==== remove diffusive domain segs from MC domain ====        
-                # drop indices from param_df
-                self._dataframe = self._dataframe.drop(mainstem_segs)
-
-                # remove keys from connections dictionary
-                for s in mainstem_segs:
-                    self._connections.pop(s)
-
-                # update downstream connections of trib segs
-                for us in trib_segs:
-                    self._connections[us] = []
 
     def create_independent_networks(self,):
 
@@ -817,121 +611,3 @@ class AbstractNetwork(ABC):
             "channel initial states complete in %s seconds."\
             % (time.time() - start_time)
         )
-
-    def build_qlateral_array(self, run, cpu_pool, supernetwork_parameters,):
-
-        start_time = time.time()
-        LOG.info("Creating a DataFrame of lateral inflow forcings ...")
-        
-        # TODO: set default/optional arguments
-        qts_subdivisions = run.get("qts_subdivisions", 1)
-        nts = run.get("nts", 1)
-        qlat_input_folder = run.get("qlat_input_folder", None)
-        qlat_input_file = run.get("qlat_input_file", None)
-
-        geo_file_type = supernetwork_parameters.get('geo_file_type')
-
-        if qlat_input_folder:
-            qlat_input_folder = pathlib.Path(qlat_input_folder)
-            if "qlat_files" in run:
-                qlat_files = run.get("qlat_files")
-                qlat_files = [qlat_input_folder.joinpath(f) for f in qlat_files]
-            elif "qlat_file_pattern_filter" in run:
-                qlat_file_pattern_filter = run.get(
-                    "qlat_file_pattern_filter", "*CHRT_OUT*"
-                )
-                qlat_files = sorted(qlat_input_folder.glob(qlat_file_pattern_filter))
-
-            qlat_file_index_col = run.get(
-                "qlat_file_index_col", "feature_id"
-            )
-
-            if geo_file_type=='NHDNetwork':
-                # Parallel reading of qlateral data from CHRTOUT
-                with Parallel(n_jobs=cpu_pool) as parallel:
-                    jobs = []
-                    for f in qlat_files:
-                        jobs.append(
-                            delayed(nhd_io.get_ql_from_chrtout)
-                            #(f, qlat_file_value_col, gw_bucket_col, terrain_ro_col)
-                            #delayed(nhd_io.get_ql_from_csv)
-                            (f)                    
-                        )
-                    ql_list = parallel(jobs)
-
-                # get feature_id from a single CHRTOUT file
-                with netCDF4.Dataset(qlat_files[0]) as ds:
-                    idx = ds.variables[qlat_file_index_col][:].filled()
-
-                # package data into a DataFrame
-                qlats_df = pd.DataFrame(
-                    np.stack(ql_list).T,
-                    index = idx,
-                    columns = range(len(qlat_files))
-                )
-
-                qlats_df = qlats_df[qlats_df.index.isin(self.segment_index)]
-            elif geo_file_type=='HYFeaturesNetwork':
-                dfs=[]
-                for f in qlat_files:
-                    df = read_file(f).set_index(['feature_id']) 
-                    dfs.append(df)
-                
-                # lateral flows [m^3/s] are stored at NEXUS points with NEXUS ids
-                nexuses_lateralflows_df = pd.concat(dfs, axis=1)  
-                
-                # Take flowpath ids entering NEXUS and replace NEXUS ids by the upstream flowpath ids
-                qlats_df = pd.concat( (nexuses_lateralflows_df.loc[int(k)].rename(v)
-                                    for k,v in self.downstream_flowpath_dict.items() ), axis=1
-                                    ).T 
-                qlats_df.columns=range(len(qlat_files))
-                qlats_df = qlats_df[qlats_df.index.isin(self.segment_index)]
-
-                # The segment_index has the full network set of segments/flowpaths. 
-                # Whereas the set of flowpaths that are downstream of nexuses is a 
-                # subset of the segment_index. Therefore, all of the segments/flowpaths
-                # that are not accounted for in the set of flowpaths downstream of
-                # nexuses need to be added to the qlateral dataframe and padded with
-                # zeros.
-                all_df = pd.DataFrame( np.zeros( (len(self.segment_index), len(qlats_df.columns)) ), index=self.segment_index,
-                    columns=qlats_df.columns )
-                all_df.loc[ qlats_df.index ] = qlats_df
-                qlats_df = all_df.sort_index()
-
-        elif qlat_input_file:
-            qlats_df = nhd_io.get_ql_from_csv(qlat_input_file)
-        else:
-            qlat_const = run.get("qlat_const", 0)
-            qlats_df = pd.DataFrame(
-                qlat_const,
-                index=self.segment_index,
-                columns=range(nts // qts_subdivisions),
-                dtype="float32",
-            )
-
-        # TODO: Make a more sophisticated date-based filter
-        max_col = 1 + nts // qts_subdivisions
-        if len(qlats_df.columns) > max_col:
-            qlats_df.drop(qlats_df.columns[max_col:], axis=1, inplace=True)
-
-        if not self.segment_index.empty:
-            qlats_df = qlats_df[qlats_df.index.isin(self.segment_index)]
-        
-        LOG.debug(
-            "lateral inflow DataFrame creation complete in %s seconds." \
-                % (time.time() - start_time)
-                )
-
-        self._qlateral = qlats_df
-
-
-
-def read_file(file_name):
-    extension = file_name.suffix
-    if extension=='.csv':
-        df = pd.read_csv(file_name)
-    elif extension=='.parquet':
-        df = pq.read_table(file_name).to_pandas().reset_index()
-        df.index.name = None
-    
-    return df
